requirements.txt
coverage==4.4.2
Flask==0.12.2
Flask-Login==0.4.0
flask-restplus==0.10.1
Flask-Cors==3.0.3
Flask-SQLAlchemy==2.3.0
psycopg2==2.7.4